refactor: rename to export_dataset
